removed Atomize module & moved safe_atom function to Recase.Generic - updated moduledoc; added atom handling  to PascalCase & ConstantCase modules - added tests
