+++ lib/recase/cases/pascal_case.ex
@@ -31,4 +31,8 @@ defmodule Recase.PascalCase do
   @spec convert(String.t()) :: String.t()
   def convert(value) when is_binary(value),
     do: rejoin(value, separator: "", case: :title)
+
+  @spec convert(atom()) :: atom()
+  def convert(value) when is_atom(value),
+    do: convert(Atom.to_string(value)) |> Recase.Generic.safe_atom()
 end
