+++ lib/recase/cases/snake_case.ex
@@ -25,5 +25,5 @@ defmodule Recase.SnakeCase do
 
   @spec convert(atom()) :: atom()
   def convert(value) when is_atom(value),
-    do: convert(Atom.to_string(value)) |> Recase.Atomize.safe_atom()
+    do: convert(Atom.to_string(value)) |> Recase.Generic.safe_atom()
 end
