+++ test/recase_test/pascal_case_test.exs
@@ -14,6 +14,12 @@ defmodule Recase.PascalCaseTest do
     assert convert("UpperCase") == "UpperCase"
   end
 
+  test "should pascal case atoms" do
+    assert convert(:upper_case) == :UpperCase
+    assert convert(:upperCase) == :UpperCase
+    assert convert(:UpperCase) == :UpperCase
+  end
+
   test "should not modify extra chars" do
     assert convert("!#$%^&*(){}[]~`'\"") == "$%^&*(){}[]~`'\""
   end
